+++ src/Psl/Internal/Loader.php
@@ -377,8 +377,9 @@ final class Loader
         'Psl\Encoding\Base64\decode',
         'Psl\Encoding\Hex\encode',
         'Psl\Encoding\Hex\decode',
-        'Psl\Shell\escape_argument',
         'Psl\Shell\execute',
+        'Psl\Shell\unpack',
+        'Psl\Shell\Internal\escape_argument',
         'Psl\Html\encode',
         'Psl\Html\encode_special_characters',
         'Psl\Html\decode',
@@ -634,6 +635,7 @@ final class Loader
         'Psl\Shell\Exception\RuntimeException',
         'Psl\Shell\Exception\PossibleAttackException',
         'Psl\Shell\Exception\TimeoutException',
+        'Psl\Shell\Exception\InvalidArgumentException',
         'Psl\Math\Exception\ArithmeticException',
         'Psl\Math\Exception\DivisionByZeroException',
         'Psl\Filesystem\Exception\RuntimeException',
@@ -743,6 +745,7 @@ final class Loader
         'Psl\Hash\Hmac\Algorithm',
         'Psl\OS\OperatingSystemFamily',
         'Psl\Password\Algorithm',
+        'Psl\Shell\ErrorOutputBehavior'
     ];
 
     public const TYPE_CONSTANTS = 1;